data/splits/split_dataset.py
"""
数据集划分脚本
将数据划分为训练、验证和测试集
"""

import os
from pathlib import Path
from sklearn.model_selection import train_test_split


def split_dataset(input_dir, output_dir, train_ratio=0.7, val_ratio=0.15, test_ratio=0.15):
    """
    划分数据集
    
    Args:
        input_dir: 输入目录
        output_dir: 输出目录
        train_ratio: 训练集比例
        val_ratio: 验证集比例
        test_ratio: 测试集比例
    """
    print(f"数据集划分 - Train: {train_ratio}, Val: {val_ratio}, Test: {test_ratio}")
    # 实现数据划分逻辑


if __name__ == '__main__':
    import argparse
    
    parser = argparse.ArgumentParser(description='数据集划分')
    parser.add_argument('--input', type=str, required=True, help='输入目录')
    parser.add_argument('--output', type=str, required=True, help='输出目录')
    parser.add_argument('--train-ratio', type=float, default=0.7, help='训练集比例')
    parser.add_argument('--val-ratio', type=float, default=0.15, help='验证集比例')
    
    args = parser.parse_args()
