scripts/calibration.py
#!/usr/bin/env python
"""
模型校准脚本
用于进行模型输出概率的校准
"""

import argparse
import torch
import yaml
from src.validation.evaluator import Evaluator


def main(args):
    """主校准函数"""
    
    print("开始模型校准...")
    print(f"模型路径: {args.model}")
    
    # 这里添加实际的校准逻辑
    # 1. 加载模型
    # 2. 加载验证数据
    # 3. 应用校准方法（如Platt scaling或Temperature scaling）


if __name__ == '__main__':
    parser = argparse.ArgumentParser(description='模型概率校准')
    parser.add_argument('--model', type=str, required=True,
                        help='模型检查点路径')
    parser.add_argument('--data-dir', type=str, default='data/',
                        help='数据目录')
    parser.add_argument('--method', type=str, default='temperature',
                        choices=['temperature', 'platt'],
                        help='校准方法')
    
    args = parser.parse_args()
    main(args)
