requirements.txt
# 深度学习框架
# 注意：macOS用户请使用 requirements-macos.txt
# 在macOS上安装PyTorch:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 机器学习和指标
scikit-learn>=1.0.0
scikit-image>=0.18.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 图像处理
Pillow>=9.0.0
opencv-python>=4.5.0

# 工具库
tqdm>=4.62.0
PyYAML>=6.0
requests>=2.27.0
pydantic>=1.8.0

# 开发工具
ipython>=7.0.0
jupyter>=1.0.0
pytest>=6.2.0

# 可选：实验追踪
# wandb>=0.12.0
# tensorboard>=2.6.0
